requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.2
Flask==3.1.3
greenlet==3.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.2.2
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
